remove list length from r match
none of our uses of the r match ever care about the length,
and the information is implicit in the list, meaning you're
consing up something just to throw it away. users who want
the length probably have to traverse the list anyway, so
they can compute the length themselves.